some more small fixes and clean up
